moved general drivers into library source directory
