+++ examples/common/ConfigApp.cpp
