+++ examples/common/ConfigApp.hpp
