+++ platformio.ini
@@ -4,7 +4,7 @@
 
 [platformio]
 src_dir = ./examples/
-lib_dir = ./
+lib_dir = ./examples/espcamlib
 
 [env]
 platform = espressif32
@@ -19,6 +19,9 @@ build_flags =
   ; -Werror
   -D CORE_DEBUG_LEVEL=0
   -D BOARD_HAS_PSRAM=1
+lib_deps =
+  nanopb/Nanopb@0.4.8
+extra_scripts = pre:prebuild.py
 
 [esp32common]
 extends = env
@@ -29,18 +32,14 @@ board_build.partitions = ./config/partitions.csv
 [env:xiao-espnow-sender]
 extends = esp32common
 board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<xiao-espnow-sender/>
-lib_deps =
-  nanopb/Nanopb@0.4.8
-  https://github.com/hpsaturn/SerialTerminal.git
-  hpsaturn/ESP32 Wifi CLI @^0.2.1
+build_src_filter = -<*> -<*common*> +<xiao-espnow-sender/>
 
 [env:freenove-tank]
 extends = esp32common
 board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<freenove-tank/>
-lib_deps = 
-  nanopb/Nanopb@0.4.8
+build_src_filter = -<*> +<common/> +<freenove-tank/>
+lib_deps =
+  ${esp32common.lib_deps}
   madhephaestus/ESP32Servo@1.1.2
   https://github.com/hpsaturn/SerialTerminal.git
   hpsaturn/ESP32 Wifi CLI @^0.2.1
@@ -48,6 +47,7 @@ lib_deps =
 [m5cores3_common]
 extends = esp32common
 lib_deps = 
+  ${esp32common.lib_deps}
   https://github.com/m5stack/M5CoreS3.git
   m5stack/M5Unified@^0.1.6
 build_flags =
@@ -59,32 +59,24 @@ build_flags =
 [env:m5cores3-espnow-sender]
 extends = m5cores3_common
 build_src_filter = -<*> +<m5cores3-espnow-sender/>
-lib_deps =
-  ${m5cores3_common.lib_deps}
-  nanopb/Nanopb@0.4.8
 
 [env:m5cores3-espnow-receiver]
 extends = m5cores3_common
 build_src_filter = -<*> +<m5cores3-espnow-receiver/>
-lib_deps =
-  ${m5cores3_common.lib_deps}
-  nanopb/Nanopb@0.4.8
 
 [env:m5core2-espnow-receiver]
 extends = esp32common
 board = esp32dev
 build_src_filter = -<*> +<m5core2-espnow-receiver/>
 lib_deps =
+  ${esp32common.lib_deps}
   m5stack/M5Unified@^0.1.6
-  nanopb/Nanopb@0.4.8
 
 [env:m5stickCplus-joystick-tank]
 extends = esp32common
 board = esp32dev
 upload_speed = 1500001
-build_src_filter = -<*> +<m5stickCplus-joystick-tank/>
+build_src_filter = -<*> +<common/> +<m5stickCplus-joystick-tank/>
 lib_deps =
+  ${esp32common.lib_deps}
   m5stack/M5Unified@^0.1.6
-  nanopb/Nanopb@0.4.8
-  https://github.com/hpsaturn/SerialTerminal.git
-  hpsaturn/ESP32 Wifi CLI @^0.2.1
